requirements.txt
asgiref==3.8.1
Django==5.1.6
django-htmx==1.22.0
djangorestframework==3.15.2
pillow==11.1.0
qrcode==5.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
